Renamed plant_modules_v2 and cleaned it up a bit
renamed plant_modules_ v2 to plant_modules and made kind of a system. also fixed imports.
