+++ Python/potential_future_build/Prototype_v2.py
@@ -1,5 +1,5 @@
 # @Date:   2021-04-21T14:03:41+02:00
-# @Last modified time: 2021-04-21T17:44:35+02:00
+# @Last modified time: 2021-04-23T12:58:40+02:00
 
 
 
@@ -8,14 +8,15 @@ This file contains codeflow suggestion for how the codeflow would work in
 the project.
 *Add more text*
 """
-from plant_modules_v2 import *
+import plant_modules
+import CoT
 import json
 
 
 # ----------------------------------------------------------------------------------------------------------------------
 
 # First time python is run (reboot system)
-plant_dictionary = plant_setup()
+plant_dictionary = plant_modules.plant_setup()
 print(plant_dictionary)
 
 while True:
@@ -23,36 +24,36 @@ while True:
     System main code to be run.
     """
     # Always checks on if user wants a new plant(create a new or switch), or save a new configuration.
-    new_plant = new_plant_configuration_key2.get()['Value']
-    save_configuration = save_configuration_key2.get()['Value']
+    new_plant = CoT.new_plant_configuration_key2.get()['Value']
+    save_configuration = CoT.save_configuration_key2.get()['Value']
 
     # Will be true if user wants to switch plant or create a new one. It will also be run if plant_dictionary does not have any dictionary
     if new_plant == 1 or plant_dictionary == 1:
-        plant_dictionary = plant_setup()
+        plant_dictionary = plant_modules.plant_setup()
 
     # Will be true if user wants to save new configuration to current plant configuration.
-    if (save_configuration == 1 and type(plant_dictionary) is dict) and (int(plant_dictionary['plant_number']) == plant_number_key2.get()['Value']):
-        plant_number = plant_number_key2.get()['Value']
-        plant_configuration(plant_number, plant_dictionary)
+    if (save_configuration == 1 and type(plant_dictionary) is dict) and (int(plant_dictionary['plant_number']) == CoT.plant_number_key2.get()['Value']):
+        plant_number = CoT.plant_number_key2.get()['Value']
+        plant_modules.plant_configuration(plant_number, plant_dictionary)
 
     # If user wanted to save configuration, but pushed on wrong plant in CoT,
     # then we'll update the signal to let user know which configuration the system is currently working with
     # and reset save_configuration.
-    elif (save_configuration == 1) and (int(plant_dictionary['plant_number']) != plant_number_key2.get()['Value']):
-            plant_number_key2.put(plant_dictionary['plant_number'])
-            save_configuration_key2.put(0)
+elif (save_configuration == 1) and (int(plant_dictionary['plant_number']) != CoT.plant_number_key2.get()['Value']):
+            CoT.plant_number_key2.put(plant_dictionary['plant_number'])
+            CoT.save_configuration_key2.put(0)
 
     #### Update sensor values ####--------------------------------------------------------------------------------------
     for plant_name in range(0, 1): # When done: range(0, len(plant_dictionary)): It wil then run trough all the plants
-        update_plant_soil_value(plant_name)
-        update_plant_water_state(plant_name)
+        plant_modules.update_plant_soil_value(plant_name)
+        plant_modules.update_plant_water_state(plant_name)
 
         #### Check sensors ####-----------------------------------------------------------------------------------------
 
-        plant_soil_check(plant_name)
+        plant_modules.plant_soil_check(plant_name)
 
         #### Water and light ####---------------------------------------------------------------------------------------
-        water(plant_name)
+        plant_modules.water(plant_name)
 
 
 
@@ -61,8 +62,8 @@ while True:
     print('PLANT STATUSES:')
     for plant_name in range(0, 2):
         print('\n  Plant', str(plant_name) +':')
-        print('  Soil:', int(plant[str(plant_name)]['soil_value']), '    Threshold:', plant[str(plant_name)]['water_requirement'])
-        print('  Pump:', plant[str(plant_name)]['water'], '  Last given water:', datetime.fromtimestamp(plant[str(plant_name)]['last_water']).strftime('%H:%M:%S %d/%m-%Y'))
+        print('  Soil:', int(plant_dictionary[str(plant_name)]['soil_value']), '    Threshold:', plant_dictionary[str(plant_name)]['water_requirement'])
+        print('  Pump:', plant_dictionary[str(plant_name)]['water'], '  Last given water:', datetime.fromtimestamp(plant_dictionary[str(plant_name)]['last_water']).strftime('%H:%M:%S %d/%m-%Y'))
 
     print('##########################################################')
 
@@ -70,4 +71,3 @@ while True:
 
 
     time.sleep(1)
-
